UPDATE slow down

+++ include/ground.h
@@ -9,7 +9,7 @@ typedef struct _Ground {
     float velocity; //Vélocité algébrique pour le déplacement des obstacles
     int type;
     int nb_obstacles; //Le nombre d'obstacles sur la ligne
-    int special_attr; //Un attribut spécial pour certains types d'obstacles (les trains :/)
+    float special_attr; //Un attribut spécial pour certains types d'obstacles (les trains :/)
 } Ground;
 
 Ground *ground_make(Obstacle **obscacles, float velocity, int type, int nb_obstacles, int special_attr);
